requirements.txt
py-cord
python-dotenv
requests
mariadb